[HDCameraEditor] Moved Camera utilities to Core
